+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Changed
 
+- Throw `IntegrationProviderAuthorizationError` on 403 response in `GraphClient`
 - Changed `azure_group_has_member` relationships from mapped to direct, because
   group members always exist in the same directory as the group.
 - Bumped `@jupiterone/integration-sdk-*@6.10.0`. This included some new required